# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FlowBB API
  version: 0.1.0
  description: |
    Kontrakt API dla prototypu FlowBB tworzonego na HackBB 2026.

    System laczy trzy moduly:
    - FLOW - wydarzenia, deklaracja uczestnictwa i plan dojazdu,
    - CREW - mikrogrupy wokol wydarzen,
    - PULSE - zagregowane dane popytu dla miasta.

    MVP nie ma pelnego uwierzytelniania. `userId` wskazuje jednego z
    seedowanych uzytkownikow demonstracyjnych.
servers:
  - url: http://localhost:5000
    description: Lokalne API podczas developmentu i prezentacji
...
paths:
...
  /api/events/{eventId}/route:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [Routing]
      summary: Zwraca proponowana trase na wydarzenie i powrot
      description: |
        Trasa zawsze jest pobierana przez `IRoutePlanner`. API zwraca
        `plannerSource: RoadRouting` dla prawdziwej trasy drogowej oraz
        `plannerSource: Demo` dla komunikacji miejskiej lub kontrolowanego fallbacku.

        Punkt startu i srodek transportu pochodza ze snapshotu deklaracji
        "Ide" (relacja `IS_GOING_TO`) zapisanego przez `upsertAttendance`.
        Klient nie przesyla wspolrzednych: backend czyta punkt startu ze
        snapshotu Attendance. Profil ani osobne pole origin nie sa ujawniane;
        mieszkaniec moze otrzymac GeoJSON swojej wyliczonej trasy. Brak
        deklaracji dla pary uzytkownik-wydarzenie oznacza 404.
      operationId: getEventRoute
      parameters:
        - $ref: "#/components/parameters/RequiredUserIdQuery"
      responses:
        "200":
          description: Plan podrozy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
              examples:
                demo:
                  value:
                    eventId: "11111111-1111-1111-1111-111111111111"
                    userId: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
                    plannerSource: Demo
                    outbound:
                      durationMinutes: 37
                      departureAt: "2026-09-19T18:12:00+02:00"
                      arrivalAt: "2026-09-19T18:49:00+02:00"
                      steps:
                        - type: Walk
                          instruction: "Idz 6 minut do przystanku."
                          durationMinutes: 6
                        - type: Transit
                          instruction: "Wsiadz do autobusu linii 7."
                          durationMinutes: 19
                          line: "7"
                        - type: Walk
                          instruction: "Idz 4 minuty na Rynek."
                          durationMinutes: 4
                    returns:
                      - durationMinutes: 34
                        departureAt: "2026-09-19T21:44:00+02:00"
                        arrivalAt: "2026-09-19T22:18:00+02:00"
                        steps:
                          - type: Transit
                            instruction: "Autobus linii 7."
                            durationMinutes: 26
                            line: "7"
                    returnGap: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      description: Id wydarzenia
      schema:
        type: string
        format: uuid
...
    RequiredUserIdQuery:
      name: userId
      in: query
      required: true
      description: Id seedowanego uzytkownika
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Nieprawidlowe dane wejsciowe
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Zasob nie istnieje
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalServerError:
      description: Nieoczekiwany blad serwera
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    PlannerSource:
      type: string
      enum: [Demo, RoadRouting, OpenTripPlanner]
    RouteGeometry:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [LineString]
        coordinates:
          type: array
          minItems: 2
          items:
            type: array
            minItems: 2
            maxItems: 2
            description: Para [longitude, latitude]; zakres latitude jest dodatkowo walidowany przez aplikacje.
            items:
              type: number
              minimum: -180
              maximum: 180
    RouteStepType:
      type: string
      enum: [Walk, Transit, Bike, Car, Wait]
    RouteStep:
      type: object
      additionalProperties: false
      required: [type, instruction, durationMinutes]
      properties:
        type:
          $ref: "#/components/schemas/RouteStepType"
        instruction:
          type: string
          maxLength: 300
        durationMinutes:
          type: integer
          minimum: 0
        line:
          type: string
          nullable: true
          maxLength: 40
    JourneyOption:
      type: object
      additionalProperties: false
      required: [durationMinutes, departureAt, arrivalAt, steps]
      properties:
        durationMinutes:
          type: integer
          minimum: 0
        distanceMeters:
          type: number
          nullable: true
          minimum: 0
        departureAt:
          type: string
          format: date-time
        arrivalAt:
          type: string
          format: date-time
        geometry:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/RouteGeometry"
        steps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RouteStep"
    RouteResponse:
      type: object
      additionalProperties: false
      required: [eventId, userId, plannerSource, outbound, returns, returnGap]
      properties:
        eventId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        plannerSource:
          $ref: "#/components/schemas/PlannerSource"
        outbound:
          $ref: "#/components/schemas/JourneyOption"
        returns:
          type: array
          items:
            $ref: "#/components/schemas/JourneyOption"
        returnGap:
          type: boolean
          description: True, gdy po wydarzeniu nie ma dogodnego powrotu
...
    ProblemDetails:
      type: object
      description: Odpowiedz zgodna z RFC 7807 / ASP.NET Core ProblemDetails
      additionalProperties: true
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          nullable: true
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        traceId:
          type: string
          nullable: true
...